remove lazysnapshot
